# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Share AI Platform.",
        "title": "Share AI Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/favorites": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前用户收藏的所有容器镜像列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "获取收藏的容器镜像列表",
                "responses": {
                    "200": {
                        "description": "data: []ContainerImage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/images": {
            "get": {
                "description": "获取所有可用的容器镜像列表，支持分页和搜索，包含镜像名称、标签、描述等信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "获取容器镜像列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（镜像名称、描述）",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: []ContainerImage, total: int",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个新的容器镜像，包括基本信息、配置参数、运行环境等详细信息",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "创建容器镜像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "镜像名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "镜像描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "镜像仓库地址",
                        "name": "registry",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "镜像仓库名称",
                        "name": "repository",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "镜像标签",
                        "name": "tag",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组织 ID",
                        "name": "org_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "可见性 (public/private)",
                        "name": "visibility",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台架构 (例如: linux/amd64)",
                        "name": "platform",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "镜像大小（字节）",
                        "name": "size",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "镜像摘要",
                        "name": "digest",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "容器暴露的端口列表",
                        "name": "ports",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "环境变量列表",
                        "name": "env",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "默认命令",
                        "name": "cmd",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "入口点命令",
                        "name": "entrypoint",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "工作目录",
                        "name": "workdir",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "运行用户",
                        "name": "user",
                        "in": "formData"
                    },
                    {
                        "type": "object",
                        "description": "镜像标签键值对",
                        "name": "labels",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "README 文件",
                        "name": "readme_file",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "标签列表",
                        "name": "tags",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.ImageResponse"
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "description": "根据镜像 ID 获取容器镜像的详细信息，包括镜像配置、版本、使用说明等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "获取容器镜像详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器镜像 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ImageResponse"
                        }
                    },
                    "404": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定容器镜像的信息，包括基本信息、配置参数等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "更新容器镜像信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器镜像 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的镜像信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ImageResponse"
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定的容器镜像及其相关配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "删除容器镜像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器镜像 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/images/{id}/collect": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "将指定的容器镜像添加到个人收藏夹中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "收藏容器镜像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器镜像 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "将指定的容器镜像从个人收藏夹中移除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-images"
                ],
                "summary": "取消收藏容器镜像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器镜像 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.ImageResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "创建者ID",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "镜像描述",
                    "type": "string"
                },
                "digest": {
                    "description": "镜像内容哈希值",
                    "type": "string"
                },
                "id": {
                    "description": "镜像唯一标识符",
                    "type": "string"
                },
                "is_starred": {
                    "description": "当前用户是否已收藏",
                    "type": "boolean"
                },
                "labels": {
                    "description": "标签列表，用于分类和搜索",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "镜像显示名称",
                    "type": "string"
                },
                "namespace": {
                    "description": "命名空间/组织",
                    "type": "string"
                },
                "org_id": {
                    "description": "组织ID",
                    "type": "string"
                },
                "platform": {
                    "description": "平台架构",
                    "type": "string"
                },
                "readme_path": {
                    "description": "README文件路径",
                    "type": "string"
                },
                "registry": {
                    "description": "镜像仓库服务器",
                    "type": "string"
                },
                "repository": {
                    "description": "镜像名称",
                    "type": "string"
                },
                "size": {
                    "description": "镜像大小（字节）",
                    "type": "integer"
                },
                "stars": {
                    "description": "收藏数",
                    "type": "integer"
                },
                "tag": {
                    "description": "版本标签",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "visibility": {
                    "description": "可见性：public/private",
                    "type": "string"
                }
            }
        },
...
        "services.UpdateImageRequest": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}